chore: move test to separate file
